add dto objects

+++ src/main/java/com/training/ecommercebackend/model/User.java
@@ -13,7 +13,8 @@ import java.util.List;
 public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
+
     @Column(name="first_name")
     private String firstName;
